requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.2
gensim==4.1.2
matplotlib==3.5.0
networkx==2.6.3
nltk==3.6.6
numpy==1.21.2
opencv_python==4.5.4.60
pandas==1.3.4
Pillow==9.0.0
protobuf==3.19.3
pydicom==2.2.2
requests==2.26.0
scikit_learn==1.0.2
seaborn==0.11.0
sentence_transformers==2.1.0
smart_open==5.2.1
spacy==3.2.0
streamlit==1.3.1
torch==1.10.1
torchsummary==1.5.1
torchtext==0.11.1
torchvision==0.11.1
tqdm==4.62.3
transformers==4.12.5